Add an icon besides percentile that pops up a modal for more information
The percentile shown on the right when the groups are sorted by contribution is confusing for people who are new to seeing the report, as no explanations on this number are given.

Let’s add an icon that, when the user clicks on it, displays a modal so that the user knows the meaning this number in detail.

A modal is a helpful visualisation that does not clutter the screen of the report that is already contained with massive amounts of information.